requirements.txt
ipython==8.18.1
grpcio==1.62.1
grpcio-tools==1.62.1
redis==5.0.3
python-dotenv==1.0.1